Commuter Patterns: building stats (#201)
Shows summary counts of building types within the selected block

Note: I changed the empty state from "Nothing selected" to "None
selected" to avoid an undesirable shift in the commuter view.

I think the shift is because the height of the text widget seems to vary
with the height of the characters _in_ the text widget. So the low
hanging tail of "g" was causing "Nothing" to render taller than "None".