Add traffic_signal_state to IntersectionSimState
There is some temporary overlap in responsibility because the
Adaptive phase type requires knowledge of the intersection state for
its waiting list, and I don't want to bring that into the new traffic
signal logic. Trying to respect Law of Demeter with the new code.

There may be something that approximates this closely enough so I can
take handling the Adaptive phase into the responsibility of the
traffic_signals module without needing this extra information.