plumb through census scenario for now
Currently this is really slow. We'll either need to speed it up
significantly or find some way to cache the scenarios and maybe
pregenerate them.

Here's a breakdown of the time spent:

    - generate census scenario took 424.5807s
      - building population areas for map took 2.8061s
	- parsing geojson took 0.4573s
	- converting to `CensusArea`s took 2.3351s
	- ... plus 0.0137s
      - assigning people to houses took 1.5885s
      - building people took 420.1861s

It's almost entirely in the `building people` step which we can
parallelize. I'm sure there are other improvements to make after that.