Reorder points when converting a PolyLine to a Polygon, so that the result is usually a Ring. This gets lanes to render in the GeoJSON export. #440
Ideally all Polygons would be expressed as an outer Ring and some
optional inner Rings. There are a few cases where this is hard, and we
instead just rely on the triangulation for rendering and point
containment and give up on GeoJSON export.

Not regenerating all maps yet -- zero screenshot diff, the map file does
change, but not sure how.