remove text and resize pregame SVG's
I wanted the SVG's to just be the image portion, and remove the baked in
text, background color, and padding.

Instead the background color, text, and padding should be applied
programmatically.

This is kind of sloppy because I did it in a text editor. Some padding
still remains since I couldn't easily translate all the points in the
svg shapes.