Handle some odd cases, where more stages were needed
In some cases oncoming left turns can't both be protected. Added checks for this and split the stage into 2. In cases where we can't reasonable work out a lagging green, an all-stages is constructed with a all-walk at the end.

Updated contains_only_crosswalks() to return a Option duration, representing the maximum amount of time. None is returned if there aren't crosswalks or there are any turns.