Issue 173 (#885)
* implemented logic to invert a polyline based on the quadrant of the polyline's angle

* Fixes for PR #885 on Issue #173.

* Reverted changes to render_curvey

Because of borrow checker limitations, I reverted the changes to the way
we call render_curvey. It is now called twice, depending on whether the
road is curved or straight. I also enclosed the implementation in an
ALWAYS-FALSE if else statement in order to deactivate it until some
remaining issues with text label rendering are resolved.

* Fix formatting

Co-authored-by: jamesneb <jnebeker@truckitapp.com>
Co-authored-by: Dustin Carlino <dabreegster@gmail.com>