Transition::Pop and title screens considered harmful
Get rid of all of the places in app startup that push on the
project-wide titlescreen onto the state stack. Instead make sure there's
a way to get to that title screen without jsut popping off all states.
Reasoning about the stack-of-states is very hard past a few special
cases.